refactor(cli): split root command enum by family
Addresses operator feedback: `src/cli/root.rs` at 413 LOC clustered
8 top-level command variants with their big help-text blocks, making
it hard to locate a specific command. Splitting by family — agent
operations and cleanup — gives each command a clearer home.

### Structural change

Each inline-args Command variant converts to a struct-wrapped
variant. The outer variant in `Command` keeps its
`#[command(before_help = BANNER, styles = HELP_STYLES, after_long_help = "...")]`
attributes; the field definitions move to an `*Args` struct with
`#[derive(Args)]` in a new sibling file.

Clap parses `Command::Load(LoadArgs)` identically to
`Command::Load { fields... }`. No CLI flag rename, no help-text
drift, no variant reorder. Verified with `cargo run -- <cmd> --help`
spot-checks on all affected subcommands.

### Split

- `src/cli/agent.rs` (226 LOC, new) — `LoadArgs`, `HardlineArgs`,
  `LaunchArgs` for interactive agent operations.
- `src/cli/cleanup.rs` (71 LOC, new) — `EjectArgs`, `PurgeArgs` for
  cleanup operations. `Exile` stays as a unit variant in
  `root.rs` (no args, no struct needed).
- `src/cli/root.rs` (201 LOC, was 413) — `Cli`, thin `Command` enum,
  3 root-level help tests.
- `src/cli/mod.rs` — adds `pub mod agent;` and `pub mod cleanup;`.

### `app/mod.rs` match-arm update

5 match arms switched from struct-variant destructuring
(`Command::Load { selector, target, ... }`) to tuple-variant with
struct destructure (`Command::Load(LoadArgs { selector, target, ... })`).
All field names unchanged; the arm bodies themselves are byte-identical
post-destructure. `Command::Exile`, `Command::Workspace { command }`,
and `Command::Config { command }` unchanged (Exile is unit; the
latter two already used wrapping for their subcommand enums).

### Test Preservation Policy

All 17 root-tests migrated, co-located with the variant they exercise:

- `agent.rs::tests` — 12 tests (`parses_load_*`, `parses_launch_*`,
  `parses_hardline_*`, plus Load/Hardline/Launch help-output tests).
- `cleanup.rs::tests` — 2 tests (`eject_help_shows_examples`,
  `purge_help_shows_examples`).
- `root.rs::tests` — 3 cross-cutting help tests
  (`root_help_shows_banner_with_top_padding`,
  `root_help_shows_all_commands`,
  `all_subcommand_help_pages_show_banner`).

The `strip_ansi` / `help_text` helpers are byte-copied into each
test module that uses them, matching the pattern Phase 9 (#135)
established for `workspace.rs` / `config.rs` test splits.

Test count: 427 → 427 (unchanged). No deletions, no weakenings.

### Visibility

Fields on each `*Args` struct are `pub` so `src/app/mod.rs` can
destructure them. No re-export at `crate::cli::*Args` — `app/mod.rs`
imports directly from `crate::cli::agent::*` and `crate::cli::cleanup::*`.
The outer enum `Command` is `pub` as before.

Zero call sites outside `src/cli/` and `src/app/mod.rs` required edits.

Verification per COMMITS.md:
- cargo fmt -- --check — pass
- cargo clippy — zero warnings
- cargo nextest run — 427 tests run: 427 passed, 0 skipped (unchanged)
- CLI help spot-check — identical output for all 6 subcommand
  help pages + root --help
